refactor: make header component
